... and another commit ...

+++ hs/ed.hs
@@ -2,7 +2,7 @@ solve :: [Int] -> [Int] -> Int
 solve [_,m,k] as = solve' k l (drop k l)
     where l = (as++take (k-1) as)
 
-softOrdering :: Int64 -> Int64 -> Ordering
+softOrdering :: Int -> Int64 -> Ordering
 softOrdering 	
 
 solve' :: Int -> [Int] -> [Int]
